Default group name to class name
With this, simulation.build_agents no longer require a redundant string
name to specify a group. It is automatically assume that the group can
be called by the classname of its constituent.

Initially, the rationale to have the explicit naming is that there could
be a use case of several groups with various params that happen to share
the same AgentClass. Even if such is the case, it is still much better
to refer the group by its Python variable name rather than its string
name.

To summarize the point: a group's Python variable name is redundant with
its string name.