Enables 'pypy' model when using @gui
When @gui is used with pypy, the gui is not displayed, but the sys.argv are read and a simulation is run from a json string.
When @gui is run in cpython with the _parameter_ pypy='pypy', then the simulation is not called in cpython, but via subprocess.call using pypy.